Удаляет мусор из кода
