+++ src/presenter/trip.js
@@ -6,26 +6,54 @@ import NoPointsView from "../view/no-points.js";
 import {RenderPosition, render, replace} from "../utils/render.js";
 import {SortType} from "../const.js";
 import {sortEventsPrice, sortEventsTime} from "../utils/utils.js";
+import {events} from "../main";
+
+export const createSortedEvents = () => {
+  const sortedEvents = [];
+  const counter = {
+    date: ``,
+    month: ``,
+    year: ``
+  };
+  for (let i = 0; i < events.length; i++) {
+    if (events[i].startDate.getFullYear() === counter.year
+      && events[i].startDate.getMonth() === counter.month
+      && events[i].startDate.getDate() === counter.date) {
+      sortedEvents[sortedEvents.length - 1].points.push(events[i]);
+    } else {
+      counter.date = events[i].startDate.getDate();
+      counter.month = events[i].startDate.getMonth();
+      counter.year = events[i].startDate.getFullYear();
+      sortedEvents.push(Object({
+        year: events[i].startDate.getFullYear(),
+        month: events[i].startDate.getMonth(),
+        date: events[i].startDate.getDate(),
+        points: [events[i]]
+      }));
+    }
+  }
+  return sortedEvents;
+};
 
 export default class Trip {
-  constructor(tripContainerComponent, arrayFromEvents) {
+  constructor(tripContainerComponent) {
     this._tripContainerComponent = tripContainerComponent;
     this._sortComponent = new SortView();
-    this._eventListComponent = new EventListView(arrayFromEvents);
+    this._sortedEvents = createSortedEvents();
+    this._eventListComponent = new EventListView(this._sortedEvents);
     this._noPointsComponent = new NoPointsView();
     this._dayContainers = this._eventListComponent.getDayContainers();
-    this._arrayFromEvents = arrayFromEvents;
     this._currentSortType = SortType.EVENT;
-    this._handleSortTypeChange = this._handleSortTypeChange.bind(this);
+    this.handleSortTypeChange = this.handleSortTypeChange.bind(this);
   }
 
   init(tripEvents) {
-    this._renderedEvents = this._arrayFromEvents;
+    this._renderedEvents = this._sortedEvents;
     this._tripEvents = tripEvents.slice();
     render(this._tripContainerComponent, this._eventListComponent, RenderPosition.BEFOREEND);
     this._renderSort();
     this._renderTripContainer();
-    this._sortComponent.setSomeMarkup(`day`);
+    this._sortComponent.setDayText(`day`);
   }
 
   _eventsSort(sortType) {
@@ -37,12 +65,12 @@ export default class Trip {
         this._renderedEvents = [{points: this._tripEvents.sort(sortEventsTime), date: ``, month: ``}];
         break;
       default:
-        this._renderedEvents = this._arrayFromEvents.slice();
+        this._renderedEvents = this._sortedEvents.slice();
     }
     this._currentSortType = sortType;
   }
 
-  _handleSortTypeChange(sortType) {
+  handleSortTypeChange(sortType) {
     if (this._currentSortType === sortType) {
       return;
     }
@@ -51,13 +79,13 @@ export default class Trip {
     this._eventListComponent = new EventListView(this._renderedEvents);
     render(this._tripContainerComponent, this._eventListComponent, RenderPosition.BEFOREEND);
     this._dayContainers = this._eventListComponent.getDayContainers();
-    this._sortComponent.setSomeMarkup(``);
+    this._sortComponent.setDayText(``);
     this._renderTripContainer();
   }
 
   _renderSort() {
     render(this._tripContainerComponent, this._sortComponent, RenderPosition.AFTERBEGIN);
-    this._sortComponent.setSortTypeChangeHandler(this._handleSortTypeChange);
+    this._sortComponent.setSortTypeChangeHandler(this.handleSortTypeChange);
   }
 
   _renderEvent(dayContainer, event) {
